Refactored update user password, and validate two factor into their own directory
